+++ tests/testthat/setup.R
@@ -12,6 +12,7 @@ vdoc <- local({
 specs <- vdoc[["specs"]]
 #  validation (F)
 
+
 #' Test harness for communication with `dv.papo`.
 #'
 #' @param mod Parameterized instance of the module to test. Should produce valid output and not trigger a `shiny::req`.
@@ -25,7 +26,7 @@ test_communication_with_papo <- function(mod, data, trigger_input_id) {
     unfiltered_dataset = datasets,
     filtered_dataset = datasets,
     module_output = function() list(),
-    utils = list(switch2 = function(id) NULL),
+    utils = list(switch2mod = function(id) NULL),
     dataset_metadata = list(name = shiny::reactive("dummy_dataset_name"))
   )